hw4/requirements.txt
sys
time
numpy
pillow
matplotlib
word2vec
nltk
sklearn
adjustText